+++ src/app/page.tsx
@@ -2,6 +2,7 @@ import { ExpertSection } from "@/components/home/ExpertSection";
 import { HeroSection } from "@/components/home/HeroSection";
 import { LearningJourneySection } from "@/components/home/LearningJourneySection";
 import { ProgramsSection } from "@/components/home/ProgramsSection";
+import { StudentOutcomesSection } from "@/components/home/StudentOutcomesSection";
 import { TrustSection } from "@/components/home/TrustSection";
 
 export default function Home() {
@@ -12,6 +13,7 @@ export default function Home() {
       <ProgramsSection />
       <LearningJourneySection />
       <ExpertSection />
+      <StudentOutcomesSection />
     </>
   );
 }